requirements.txt
# Core
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6

# OpenML
openml==0.14.0

# Meta-features
pymfe==0.4.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Progress bars
tqdm==4.65.0
joblib==1.3.1

# Configuration
pyyaml==6.0
python-dotenv==1.0.0

# Jupyter
jupyter==1.0.0
ipykernel==6.25.0

# Testing
pytest==7.4.0